extract utils depending on Blockly to their own file
