+++ apps/src/sites/studio/pages/embedBlocks.js
@@ -2,7 +2,7 @@
  * Entry point to build a bundle containing a set of globals used when displaying
  * embedded blocks via blockly-in-an-iframe.
  */
-import {convertXmlToBlockly} from '@cdo/apps/templates/instructions/utils';
+import {convertXmlToBlockly} from '@cdo/apps/util/blocklyUtils';
 
 window.addEventListener('DOMContentLoaded', () => {
   convertXmlToBlockly(document.body);
